+++ benchmark/src/infra.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 use testcontainers::{ContainerAsync, GenericImage, ImageExt, core::WaitFor, runners::AsyncRunner};
 
 // ============================================================
-// Postgres with benchmark-tuned CLI flags (bypasses default healthcheck)
+// Postgres with benchmark-tuned CLI flags (bypasses default health check)
 
 const PG_CMD: [&str; 13] = [
     "postgres",
@@ -80,25 +80,6 @@ impl InfraStack {
             redis_container,
         })
     }
-
-    // ============================================================
-    // Needs to be in the orchestrator ...
-
-    // /// Returns a pre-configured `Command` builder for the `lobby` binary.
-    // /// Injects dynamic URLs, disables source `.env` requirement, and sets benchmark flags.
-    // pub fn lobby_command(&self, api_keys: Vec<ApiStack>) -> Command {
-    //     let mut cmd = Command::new("cargo");
-    //     cmd.args(["run", "--release", "--features", "bench", "--bin", "lobby"])
-    //         .env("DATABASE_URL", &self.pg_url)
-    //         .env("REDIS_URL", &self.redis_url)
-    //         .env("SERVER_ADDR", "127.0.0.1:3000")
-    //         .env("RUST_LOG", "WARN"); // Reduce overhead during bench
-
-    //     for api_key in api_keys {
-    //         cmd.env(api_key.key, api_key.api);
-    //     }
-    //     cmd
-    // }
 
     // ============================================================
