+++ benchmark/src/metrics/deserialize.rs
