+++ crate/cortex/src/artifacts/mod.rs
@@ -3,8 +3,6 @@ pub mod error;
 pub mod pool;
 pub mod retry;
 pub mod state;
-#[cfg(feature = "benchmark-telemetry")]
-pub mod telemetry;
 
 // ============================================================
 // re-exports
@@ -14,7 +12,5 @@ pub use error::*;
 pub use pool::*;
 pub use retry::*;
 pub use state::*;
-#[cfg(feature = "benchmark-telemetry")]
-pub use telemetry::*;
 
 // ============================================================
